pyproject.toml
[tool.poetry]
name = "pykoi"
version = "0.0.10"
description = "pykoi: Active learning in one unified interface"
authors = ["CambioML <wanwanaiai45@gmail.com>"]
license = "Apache"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
packaging = "23.1"
fastapi = "0.100.0"
pydantic = "2.5.2"
starlette = "0.27.0"
uvicorn = "0.23.1"
scipy = "1.11.1"
openai = "1.6.1"
passlib = "1.7.4"
bcrypt = "4.0.1"
posthog = "3.0.1"
pynvml = "11.5.0"
pandas = "2.0.3"
python-dotenv = "^1.0.0"

transformers = { version = "4.36.2", optional = true }
einops = { version = "0.6.1", optional = true }
accelerate = { version = "0.21.0", optional = true }
bitsandbytes = { version = "0.40.2", optional = true }

langchain = { version = "0.0.338", optional = true }
scikit-learn = { version = "1.3.0", optional = true }
chromadb = { version = "0.4.20", optional = true }
pyepsilla = { version = ">=0.1.1", optional = true }
pdfminer-six = { version = "20221105", optional = true }
docx2txt = { version = "0.8", optional = true }
python-multipart = { version = "0.0.6", optional = true }
tiktoken = { version = "0.4.0", optional = true }
sentence-transformers = { version = "2.2.2", optional = true }

datasets = { version = "2.15.0", optional = true }
evaluate = { version = "0.4.0", optional = true }
peft = { version = "0.7.1", optional = true }
trl = { version = "0.7.4", optional = true }

[tool.poetry.extras]
huggingface = [
    "transformers",
    "einops",
    "accelerate",
    "bitsandbytes",
    "sentence-transformers",
    "xformers"
]
rag = [
    "langchain",
    "scikit-learn",
    "chromadb",
    "pyepsilla",
    "pdfminer-six",
    "docx2txt",
    "python-multipart",
    "tiktoken"
]
rlhf = [
    "scikit-learn",
    "transformers",
    "einops",
    "accelerate",
    "bitsandbytes",
    "datasets",
    "evaluate",
    "peft",
    "trl"
]

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
